+++ Tests/ricaun.Revit.Templates.Tests/Utils/CmdRunner.cs
@@ -1,0 +1,48 @@
+using System;
+using System.Diagnostics;
+using System.Threading;
+using System.Threading.Tasks;
+
+namespace ricaun.Revit.Templates.Tests.Utils;
+
+public class CmdRunner
+{
+    public static async Task<int> RunAsync(string fileName, params string[] arguments)
+    {
+        using var process = new Process
+        {
+            StartInfo = new ProcessStartInfo
+            {
+                FileName = fileName,
+                Arguments = $"{string.Join(" ", arguments)}",
+                UseShellExecute = false,
+                RedirectStandardOutput = true,
+                RedirectStandardError = true,
+                StandardOutputEncoding = System.Text.Encoding.UTF8,
+                StandardErrorEncoding = System.Text.Encoding.UTF8,
+                CreateNoWindow = true
+            }
+        };
+
+        process.OutputDataReceived += (_, e) =>
+        {
+            if (e.Data != null)
+                Console.WriteLine(e.Data);
+        };
+
+        process.ErrorDataReceived += (_, e) =>
+        {
+            if (e.Data != null)
+                Console.Error.WriteLine(e.Data);
+        };
+
+        process.Start();
+
+        process.BeginOutputReadLine();
+        process.BeginErrorReadLine();
+
+        process.WaitForExit(120000);
+
+        return process.ExitCode;
+    }
+}